clean up test organization
